marketplace/requirements.txt
asgiref==3.6.0
black==23.1.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
Django==3.2.18
django-clearcache==1.2.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.4.1
django-sql-utils==0.6.1
filelock==3.10.4
identify==2.5.22
isort==5.12.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pre-commit==3.2.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
typing_extensions==4.5.0
virtualenv==20.21.0
celery==5.2.6
redis==4.5.4
requests==2.28.2
django-mathfilters==1.0.0
django-admin-interface==0.26.0
python-dotenv==1.0.0